Added Electron support for MacOS.
