+++ package.json
@@ -3,13 +3,20 @@
   "version": "0.1.1",
   "description": "Get active window title.",
   "main": "index.js",
-  "contributors":[{ "name" : "Jason Stallings "
-                  , "email" : "jason@stallin.gs"
-                  , "url" : ""},
-                  { "name" : "Robson Vieira"
-                  , "email" : "robsonvnasc@gmail.com"
-                  , "url" : ""}
-                ],
+  "contributors": [
+    {
+      "name": "Jason Stallings ",
+      "email": "jason@stallin.gs",
+      "url": ""
+    },
+    {
+      "name": "Robson Vieira",
+      "email": "robsonvnasc@gmail.com",
+      "url": ""
+    }
+  ],
   "license": "MIT",
-  "dependencies": {}
+  "dependencies": {
+    "node-osascript": "1.0.4"
+  }
 }
